Data-Pipeline/requirements-airflow.txt
# Core Airflow
apache-airflow==2.9.2
psycopg2-binary==2.9.9

# Transcription & ML (for scripts)
faster-whisper==1.2.0
sentence-transformers==5.1.1
transformers==4.57.1
torch==2.9.0
torchaudio==2.9.0
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.7.2

# Utilities
requests==2.32.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2

# GCP (for DVC and storage)
google-cloud-storage==3.4.1
google-cloud-core==2.4.3
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-api-core==2.27.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0

# DVC (for data versioning)
dvc==3.63.0
dvc-gs==3.0.2
dvc-data==3.16.12
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvc-http==2.32.0
dulwich==0.24.7
scmrepo==3.5.2
sqltrie==0.11.2

# Validation & Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
jiwer==4.0.0
rouge-score==0.1.2
requests-mock==1.12.1

# Data processing
pandas==2.3.3

# Dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
packaging==25.0
PyYAML==6.0.3
click==8.3.0
tqdm==4.67.1
filelock==3.20.0
fsspec==2025.9.0
gcsfs==2025.9.0
cloudpathlib==0.23.0

# Hugging Face
huggingface-hub==0.35.3
tokenizers==0.22.1
safetensors==0.6.2

# System
psutil==7.1.1
distro==1.9.0

